nooo

+++ BitSimulator/BitSimulator/Utilities.cpp
@@ -23,9 +23,9 @@ void ChangeColor(sf::Image * ptr, sf::Color col1, sf::Color col2)
 {
 	Vector2u dim = ptr->getSize();
 
-	for (int posx = 0; posx < dim.x; posx++)
+	for (unsigned posx = 0; posx < dim.x; posx++)
 	{
-		for (int posy = 0; posy < dim.y; posy++)
+		for (unsigned posy = 0; posy < dim.y; posy++)
 		{
 			if (ptr->getPixel(posx, posy) == col1)
 				ptr->setPixel(posx, posy, col2);
@@ -48,22 +48,22 @@ double CoverBonus(sf::FloatRect elem1, sf::FloatRect elem2)
 
 double CoverPercentage(sf::FloatRect elem1, sf::FloatRect elem2)
 {
-		int area = elem1.width * elem1.height;
+		int area = (int)(elem1.width * elem1.height);
 
-		int diff_x = elem1.left - elem2.left;
+		int diff_x = (int)(elem1.left - elem2.left);
 
 		if (diff_x < 0)
-			diff_x = elem1.width + diff_x;
+			diff_x =(int)(elem1.width + diff_x);
 
 		else
-			diff_x = elem2.width - diff_x;
+			diff_x = (int)(elem2.width - diff_x);
 
-		int diff_y = elem1.top - elem2.top;
+		int diff_y = (int)(elem1.top - elem2.top);
 
 		if (diff_y < 0)
-			diff_y = elem1.height + diff_y;
+			diff_y = (int)(elem1.height + diff_y);
 		else
-			diff_y = elem2.height - diff_y;
+			diff_y = (int)(elem2.height - diff_y);
 		
 		double l = (double)(diff_x * diff_y) / area;
 		return l;
